+++ tests/04-memory/src/Main.h
@@ -1,10 +1,8 @@
 #ifndef MAIN_H
 #define MAIN_H
 
-#include <purescript.h>
+#include "runtime/purescript.h"
 
-PURS_FFI_FUNC_2(Main_sub, x, y, {
-	return purs_any_int(purs_any_get_int(x) - purs_any_get_int(y));
-});
+PURS_FFI_EXPORT(Main_sub);
 
 #endif // MAIN_H
